Fixed package locations
