+++ backend/src/FECommunication.java
@@ -1,5 +1,6 @@
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import jsonObjects.Student;
 
 public class FECommunication {
